Modify SetLocation method
